Uncertainty of KL divergence estimate using Efron's bootstrap.

This function computes a confidence interval for KL divergence based on Efron's bootstrap. The approach only works for kernel density-based estimators since nearest neighbour-based estimators cannot deal with the ties produced when sampling with replacement.

Usage

kld_ci_bootstrap(X, Y, estimator = kld_est_kde1, B = 500L, alpha = 0.05)

Arguments

X, Y

n-by-d and m-by-d matrices, representing n samples from the true distribution \(P\) and m samples from the approximate distribution \(Q\), both in d dimensions. Vector input is treated as a column matrix.

estimator

A function expecting two inputs X and Y, the Kullback-Leibler divergence estimation method. Defaults to kld_est_kde1, which can only deal with one-dimensional two-sample problems (i.e., d = 1 and q = NULL).

B

Number of bootstrap replicates (default: 500), the larger, the more accurate, but also more computationally expensive.

alpha

Error level, defaults to 0.05.

Value

A list with the fields "est" (the estimated KL divergence), "boot" (a length B numeric vector with KL divergence estimates on the bootstrap samples), and "ci" (a length 2 vector containing the lower and upper limits of the estimated confidence interval).

Details

Reference:

Efron, "Bootstrap Methods: Another Look at the Jackknife", The Annals of Statistics, Vol. 7, No. 1 (1979).

Examples

# 1D Gaussian, two samples
X <- rnorm(100)
Y <- rnorm(100, mean = 1, sd = 2)
kld_gaussian(mu1 = 0, sigma1 = 1, mu2 = 1, sigma2 = 2^2)
#> [1] 0.4431472
kld_est_kde1(X, Y)
#> [1] 0.4900294
kld_ci_bootstrap(X, Y)
#> $est
#> [1] 0.4900294
#> 
#> $boot
#>   [1] 0.5864032 0.4608133 0.4571713 0.5552641 0.4580054 0.3607945 0.5542766
#>   [8] 0.7211304 0.4283980 0.5340580 0.5177979 0.5511385 0.5065323 0.3865519
#>  [15] 0.4298690 0.5065186 0.5087559 0.4492978 0.5933967 0.5071220 0.6054198
#>  [22] 0.5218433 0.5144037 0.7139447 0.5341179 0.5143128 0.4956566 0.5334696
#>  [29] 0.6336772 0.5903280 0.6460193 0.5498371 0.7043385 0.5591346 0.3838133
#>  [36] 0.6805425 0.5469331 0.5626581 0.4431756 0.6224077 0.6512727 0.4304299
#>  [43] 0.3999499 0.4229542 0.4567004 0.4602692 0.5126987 0.8204335 0.5959872
#>  [50] 0.6104068 0.6389173 0.5608872 0.5580139 0.4691429 0.4630768 0.3648030
#>  [57] 0.5539104 0.4431243 0.5301480 0.4575464 0.5889774 0.4186662 0.4746168
#>  [64] 0.4910354 0.4244641 0.5066606 0.6653451 0.5976784 0.5314204 0.3463020
#>  [71] 0.7157521 0.4146493 0.4622089 0.5974708 0.5136973 0.3790035 0.4861971
#>  [78] 0.4713702 0.5129981 0.4444806 0.5488354 0.3888370 0.5326723 0.3877294
#>  [85] 0.3805600 0.5290375 0.5498235 0.4462557 0.6304899 0.3643159 0.5292193
#>  [92] 0.5237657 0.5809993 0.4581182 0.5000498 0.4029632 0.6962937 0.5138156
#>  [99] 0.4620032 0.5181823 0.6366196 0.4798599 0.5163507 0.5715622 0.5388536
#> [106] 0.3923992 0.5632007 0.7137795 0.4171165 0.3316600 0.6276278 0.5590074
#> [113] 0.5245280 0.4863687 0.4288414 0.3128432 0.4159403 0.4568610 0.5029520
#> [120] 0.4864586 0.3959727 0.5428290 0.5827779 0.6143517 0.5826319 0.4833914
#> [127] 0.4412014 0.4241867 0.4053320 0.3669919 0.4533211 0.6354178 0.5197687
#> [134] 0.4700752 0.5167969 0.5541567 0.5391346 0.5888428 0.5850539 0.4417550
#> [141] 0.3976842 0.5263029 0.5377313 0.4128433 0.4987431 0.4077401 0.6882617
#> [148] 0.4840892 0.6520077 0.5730685 0.6614839 0.6014170 0.5322079 0.6159163
#> [155] 0.7344208 0.5206943 0.7204053 0.4093092 0.4574874 0.6067283 0.4342283
#> [162] 0.4726482 0.4041033 0.3608052 0.5372257 0.4420284 0.6444240 0.3677563
#> [169] 0.4073221 0.6494383 0.6451980 0.4571245 0.6303185 0.5275814 0.5170395
#> [176] 0.5490211 0.7134083 0.6147787 0.4877667 0.6420895 0.3244239 0.7761291
#> [183] 0.6285949 0.5926184 0.3147137 0.3962720 0.3362052 0.4430977 0.6537422
#> [190] 0.6673896 0.3982999 0.4463706 0.4635846 0.5473231 0.5694183 0.4457561
#> [197] 0.4746200 0.6429025 0.5910925 0.3637832 0.5263520 0.4432945 0.4628161
#> [204] 0.3630364 0.5500146 0.5953389 0.3609131 0.5759473 0.5504926 0.4041231
#> [211] 0.6526427 0.4207580 0.5703979 0.4770280 0.4014190 0.4819902 0.5321195
#> [218] 0.6003962 0.6080038 0.4867250 0.5567113 0.4509496 0.4284230 0.4099115
#> [225] 0.4510461 0.7552816 0.5725751 0.5117585 0.4779506 0.5399530 0.5543283
#> [232] 0.4553394 0.6132559 0.5750282 0.3652938 0.6121478 0.4026437 0.4203638
#> [239] 0.4911227 0.5256492 0.4870966 0.5918262 0.5278378 0.3521379 0.5410300
#> [246] 0.4663112 0.6259421 0.7201454 0.5315399 0.5684498 0.6106262 0.4809707
#> [253] 0.4950502 0.5440829 0.3733618 0.6189908 0.6092458 0.4026438 0.5177402
#> [260] 0.5636568 0.5305590 0.6233414 0.5183211 0.5313470 0.6125496 0.6057815
#> [267] 0.4863786 0.5355574 0.4679212 0.5233997 0.4760461 0.5353631 0.4522979
#> [274] 0.4267231 0.3696970 0.3522227 0.4332185 0.5609712 0.6021262 0.4183028
#> [281] 0.4456215 0.4567423 0.5693874 0.5866731 0.5073995 0.4330396 0.4964139
#> [288] 0.4885557 0.6123405 0.5861728 0.4484435 0.4941168 0.6846272 0.3652634
#> [295] 0.5960890 0.6338155 0.5343186 0.5786262 0.3592598 0.9809490 0.3073490
#> [302] 0.4464713 0.4682712 0.4481374 0.6090096 0.5667751 0.5267679 0.5988622
#> [309] 0.5095167 0.5407229 0.6180079 0.5154980 0.3239159 0.5582658 0.4012783
#> [316] 0.5544350 0.6041993 0.7958377 0.5466469 0.5811071 0.4909092 0.4815430
#> [323] 0.4298053 0.5942197 0.4812544 0.4926729 0.6015869 0.5402516 0.3224224
#> [330] 0.4687033 0.4313008 0.5752823 0.5085402 0.5812202 0.3789923 0.4281671
#> [337] 0.4497657 0.5964637 0.4659661 0.4827378 0.5392515 0.5610209 0.5398243
#> [344] 0.2983453 0.4872457 0.4752025 0.6206957 0.5537890 0.5186084 0.4166311
#> [351] 0.7002251 0.7968110 0.5664924 0.6439979 0.6382467 0.5959192 0.3883742
#> [358] 0.4363849 0.4811549 0.5146107 0.6615987 0.5648167 0.5018244 0.6422222
#> [365] 0.5525348 0.4956365 0.5819250 0.6477299 0.4731876 0.5983845 0.5642013
#> [372] 0.4856003 0.5328077 0.5072100 0.6051188 0.3935036 0.5215499 0.5376027
#> [379] 0.4611302 0.6995328 0.8873190 0.3819835 0.5735582 0.6096235 0.7041087
#> [386] 0.3984228 0.5070268 0.3859903 0.4934464 0.5008449 0.5898929 0.4913278
#> [393] 0.6338764 0.3378219 0.4868090 0.3502979 0.3891741 0.6108247 0.5449756
#> [400] 0.5033959 0.5961160 0.6798923 0.5210183 0.4420096 0.4480456 0.6173787
#> [407] 0.4511347 0.3955482 0.4802380 0.7068094 0.3743767 0.6386850 0.7219368
#> [414] 0.6046990 0.4496600 0.5156298 0.5376671 0.6530752 0.5139836 0.6548601
#> [421] 0.5445581 0.5508575 0.5568319 0.4620795 0.3002349 0.6468904 0.5921987
#> [428] 0.5610357 0.5154120 0.6116259 0.4812518 0.5645279 0.5042029 0.5208802
#> [435] 0.4384623 0.4952115 0.4101180 0.5029570 0.5153480 0.5274779 0.4327123
#> [442] 0.4906283 0.4926505 0.4251515 0.4887082 0.3544052 0.8385778 0.4661866
#> [449] 0.4778191 0.4937673 0.5704667 0.6212294 0.8944052 0.5639983 0.4285845
#> [456] 0.3887443 0.5239088 0.5370626 0.6517257 0.5054571 0.5232731 0.6101302
#> [463] 0.4995656 0.4669690 0.5455549 0.5311537 0.3153705 0.4729579 0.4552241
#> [470] 0.6009929 0.6979059 0.4184252 0.5717298 0.5504107 0.7063055 0.5763016
#> [477] 0.4306410 0.5542793 0.5639440 0.4900803 0.3696576 0.3960363 0.5441732
#> [484] 0.6220006 0.4945953 0.5953638 0.3676641 0.5667015 0.4864452 0.4826337
#> [491] 0.4233900 0.4129457 0.4728503 0.4658586 0.8295637 0.4405745 0.5383343
#> [498] 0.4511211 0.7120132 0.5368383
#> 
#> $ci
#>     97.5%      2.5% 
#> 0.2592728 0.6318588 
#>