Uncertainty of KL divergence estimate using Efron's bootstrap.

This function computes a confidence interval for KL divergence based on Efron's bootstrap. The approach only works for kernel density-based estimators since nearest neighbour-based estimators cannot deal with the ties produced when sampling with replacement.

Usage

kld_ci_bootstrap(X, Y, estimator = kld_est_kde1, B = 500L, alpha = 0.05)

Arguments

X, Y

n-by-d and m-by-d matrices, representing n samples from the true distribution \(P\) and m samples from the approximate distribution \(Q\), both in d dimensions. Vector input is treated as a column matrix.

estimator

A function expecting two inputs X and Y, the Kullback-Leibler divergence estimation method. Defaults to kld_est_kde1, which can only deal with one-dimensional two-sample problems (i.e., d = 1 and q = NULL).

B

Number of bootstrap replicates (default: 500), the larger, the more accurate, but also more computationally expensive.

alpha

Error level, defaults to 0.05.

Value

A list with the fields "est" (the estimated KL divergence), "boot" (a length B numeric vector with KL divergence estimates on the bootstrap samples), and "ci" (a length 2 vector containing the lower and upper limits of the estimated confidence interval).

Details

Reference:

Efron, "Bootstrap Methods: Another Look at the Jackknife", The Annals of Statistics, Vol. 7, No. 1 (1979).

Examples

# 1D Gaussian, two samples
X <- rnorm(100)
Y <- rnorm(100, mean = 1, sd = 2)
kld_gaussian(mu1 = 0, sigma1 = 1, mu2 = 1, sigma2 = 2^2)
#> [1] 0.4431472
kld_est_kde1(X, Y)
#> [1] 0.7355116
kld_ci_bootstrap(X, Y)
#> $est
#> [1] 0.7355116
#> 
#> $boot
#>   [1] 0.8343221 0.7508969 0.6398676 0.6143932 0.7256303 0.7141915 0.7489336
#>   [8] 0.7757240 0.6512059 0.7919391 0.7942396 0.8931893 0.8922686 0.9217720
#>  [15] 0.7699415 0.8543307 0.7052056 0.8402660 0.6342571 0.6876270 0.8930348
#>  [22] 0.8204401 0.9027163 0.6954667 0.9086848 0.6318139 0.7148039 0.7279260
#>  [29] 0.8568061 0.7522284 0.6937782 0.8512711 0.6303072 0.5915244 0.9710108
#>  [36] 0.7854252 0.8765962 0.8032517 0.6456086 0.5735262 0.8893643 0.8087910
#>  [43] 0.7936836 0.5747577 0.8140469 0.7917568 0.8043973 0.7316396 0.5681966
#>  [50] 0.8718564 0.6773901 0.7642510 0.8010287 0.8754750 0.7572071 0.8656637
#>  [57] 0.7418016 0.6780003 1.1526167 0.8667904 0.6890156 0.9652416 0.6616116
#>  [64] 0.6896635 0.8906805 0.7802338 0.8808687 0.7068650 0.7373339 0.8172529
#>  [71] 0.8259557 0.6915695 0.7787341 0.8536835 0.6149805 0.7873545 0.7397041
#>  [78] 0.6055451 0.7570900 0.7588512 0.5310581 0.8302364 0.8695659 0.7538020
#>  [85] 0.5810937 0.9528992 0.8178780 0.8500075 0.8480525 0.7097118 1.0210138
#>  [92] 0.8023033 0.8968113 0.7333741 0.8024978 0.8520625 0.7596676 0.7575825
#>  [99] 0.7349362 0.6068280 0.6995539 0.6484275 0.7659140 0.9190852 0.7241748
#> [106] 0.9394559 0.8999372 0.8710466 0.7100967 0.9756273 0.8641662 0.7642985
#> [113] 0.6992259 0.7006891 0.6973892 0.7978066 0.6467821 0.6745370 0.7513538
#> [120] 0.8694410 0.7120030 0.8767314 0.9074854 0.5087223 0.7085607 0.7797679
#> [127] 0.8125687 0.8258917 0.7851355 0.6574712 0.6992928 0.8698369 0.7906002
#> [134] 0.8842903 0.7028842 0.7898644 0.7256250 1.0481949 0.9406040 0.7791016
#> [141] 0.8750820 0.9962454 0.8230295 0.6863598 0.9331050 0.5879106 1.0030768
#> [148] 0.6050053 0.8687599 0.7577799 0.7682543 0.5924897 0.7721508 0.5342687
#> [155] 1.0090220 0.8917448 0.8178776 0.9434399 0.8005488 0.8687898 0.5930912
#> [162] 0.8284728 0.8414345 0.7434748 0.6858792 0.6669080 0.7498117 0.7671221
#> [169] 0.9712774 0.7587813 0.6849698 0.7184792 0.5470174 0.6200220 0.8267181
#> [176] 0.8637986 0.6525368 0.6977343 0.7752001 0.7977158 0.9187862 0.6005723
#> [183] 0.9434298 0.8964728 0.7093990 0.9232413 0.9099688 0.6929191 0.7790669
#> [190] 0.7247702 0.9392681 0.8956784 0.9286884 0.8432006 0.7765655 0.7460071
#> [197] 0.8527312 0.9162530 0.7458499 0.9049660 1.0196240 0.6360945 0.7053549
#> [204] 0.8329185 0.7295061 0.7222708 0.8145204 0.8443866 0.7746705 0.7297948
#> [211] 0.7069799 1.0587692 0.7749457 0.7832796 0.6042718 0.5482808 0.7179587
#> [218] 0.6603117 0.6479836 0.7135213 1.0189800 0.6678817 0.7618408 0.7840206
#> [225] 0.7290790 0.6402851 0.6871964 0.8735579 0.7745316 0.9023459 0.8103658
#> [232] 0.8137962 0.7490958 0.8117469 0.6113596 0.8769631 0.7208511 1.0069166
#> [239] 0.5514560 0.8723149 0.8035313 0.8297648 0.6172196 0.8363986 0.7653939
#> [246] 0.6676829 1.0122747 0.6530408 0.7024819 1.0169991 0.7958981 0.7052096
#> [253] 0.7249546 0.7162652 0.8600716 0.9837453 0.7404046 0.7988204 0.8770446
#> [260] 0.8596873 0.7346405 1.0228906 0.8942483 0.5959080 0.8316394 0.9428461
#> [267] 0.8001280 0.7753452 0.6761619 0.8241252 0.6879136 0.7197313 0.9346920
#> [274] 0.8899650 0.9503829 0.8404767 0.8674541 0.9218359 0.6804084 0.6433382
#> [281] 0.9331106 0.7356232 0.7914607 0.7602190 0.6462540 0.7291505 0.6869131
#> [288] 0.6832913 0.7436355 0.7065148 0.7587640 0.7810815 0.8849286 0.5197993
#> [295] 0.7623273 0.6366240 0.9802509 0.7866199 0.7877852 0.6393252 0.7814666
#> [302] 0.8636214 0.8234177 0.8343129 0.7158334 0.8927471 0.7590852 1.0544650
#> [309] 0.6858445 0.6544542 0.8747150 0.8676164 0.7490985 0.7121611 0.6598107
#> [316] 0.7317064 0.7313981 0.8786487 0.8671184 0.6885050 0.7305042 0.7734092
#> [323] 0.8594598 1.0539967 0.7813297 0.7402808 0.7463684 0.7956104 0.7484062
#> [330] 0.8056472 0.5132838 1.0328820 0.6747563 0.9633426 0.8031103 0.7695461
#> [337] 0.9444136 0.9418736 0.8123439 0.6467675 0.7226571 0.9178119 0.6620231
#> [344] 0.7608322 0.7403128 0.9747221 0.7095665 0.8610613 0.7633247 0.7970302
#> [351] 0.9416030 0.9412531 0.7607892 0.6885227 0.8286359 0.5830347 0.8231205
#> [358] 0.8378931 0.8062200 0.9951659 0.8099558 0.6244166 0.6973802 0.8241677
#> [365] 0.8890443 0.6731018 0.7778948 0.5389394 0.5646960 0.8871744 0.6786931
#> [372] 0.9708098 0.8483174 0.6975821 0.7328943 0.7679017 0.9122534 0.7175540
#> [379] 0.5866442 0.9482200 0.8273716 0.8009590 0.8456013 0.7306357 0.9055317
#> [386] 0.8353771 0.6432551 0.9648899 0.8578815 0.7517897 0.6031693 0.6620598
#> [393] 0.9446988 0.9773091 0.6467486 0.7123060 0.7583915 0.6337573 0.7437617
#> [400] 0.6782834 0.8327958 0.7769548 0.8550338 0.7080330 0.7150708 0.7736996
#> [407] 0.8578995 0.5811065 0.8664577 0.6993159 0.7135191 0.8904868 0.9655642
#> [414] 0.8109373 0.8915709 0.5735860 0.8166857 0.6080086 0.8830728 0.9372066
#> [421] 0.8118337 0.6782216 0.7785828 0.7090160 0.7263374 0.6257635 0.6581204
#> [428] 0.7035676 0.8359827 0.7182365 0.7761743 0.5496484 0.6326138 0.7463562
#> [435] 0.7223872 0.7881124 0.8393770 0.5779815 0.7116884 0.7172485 0.7851067
#> [442] 0.8186132 0.8613520 0.7799489 0.7633363 0.7759348 0.9212215 0.7330316
#> [449] 0.7033022 0.9351550 0.8014444 0.9775444 0.8949554 0.7856337 0.8016677
#> [456] 0.6622553 0.7468117 0.6905639 0.6442310 0.5797946 0.7434273 0.9101827
#> [463] 0.7307969 0.8301657 0.8799418 0.8062991 1.0656158 1.0676531 0.6748731
#> [470] 0.9143507 0.7482958 0.6447234 0.6017000 0.9791848 0.6762234 0.6290499
#> [477] 0.9411008 0.8983670 0.8677831 0.7927996 0.7478380 0.8328756 0.6337062
#> [484] 0.9531542 0.6423541 0.8495404 1.0292958 0.7393628 0.9337021 0.7482195
#> [491] 0.5913438 0.8781677 0.9338889 0.8574773 0.8502059 0.8172465 0.9096540
#> [498] 0.6679631 0.9062599 0.8773872
#> 
#> $ci
#>     97.5%      2.5% 
#> 0.4529842 0.8974686 
#> 
#> $se
#> [1] 0.114973
#>