Uncertainty of KL divergence estimate using Efron's bootstrap.

This function computes a confidence interval for KL divergence based on Efron's bootstrap. The approach only works for kernel density-based estimators since nearest neighbour-based estimators cannot deal with the ties produced when sampling with replacement.

Usage

kld_ci_bootstrap(X, Y, estimator = kld_est_kde1, B = 500L, alpha = 0.05)

Arguments

X, Y

n-by-d and m-by-d matrices, representing n samples from the true distribution \(P\) and m samples from the approximate distribution \(Q\), both in d dimensions. Vector input is treated as a column matrix.

estimator

A function expecting two inputs X and Y, the Kullback-Leibler divergence estimation method. Defaults to kld_est_kde1, which can only deal with one-dimensional two-sample problems (i.e., d = 1 and q = NULL).

B

Number of bootstrap replicates (default: 500), the larger, the more accurate, but also more computationally expensive.

alpha

Error level, defaults to 0.05.

Value

A list with the fields "est" (the estimated KL divergence), "boot" (a length B numeric vector with KL divergence estimates on the bootstrap samples), and "ci" (a length 2 vector containing the lower and upper limits of the estimated confidence interval).

Details

Reference:

Efron, "Bootstrap Methods: Another Look at the Jackknife", The Annals of Statistics, Vol. 7, No. 1 (1979).

Examples

# 1D Gaussian, two samples
X <- rnorm(100)
Y <- rnorm(100, mean = 1, sd = 2)
kld_gaussian(mu1 = 0, sigma1 = 1, mu2 = 1, sigma2 = 2^2)
#> [1] 0.4431472
kld_est_kde1(X, Y)
#> [1] 0.5336657
kld_ci_bootstrap(X, Y)
#> $est
#> [1] 0.5336657
#> 
#> $boot
#>   [1] 0.5736891 0.5358512 0.4749976 0.3882043 0.6145555 0.6550352 0.4359620
#>   [8] 0.7509436 0.6154062 0.7187383 0.6203521 0.6862976 0.5953209 0.6449310
#>  [15] 0.6319719 0.5509055 0.6975013 0.8439371 0.5432705 0.7253547 0.5985104
#>  [22] 0.5929938 0.6073017 0.5219260 0.5775205 0.7065137 0.4179322 0.6425269
#>  [29] 0.3916792 0.3947691 0.4752119 0.6078384 0.4683926 0.5275175 0.6142741
#>  [36] 0.5029576 0.5592827 0.4906330 0.6368965 0.6307913 0.7014355 0.6197439
#>  [43] 0.5582100 0.5955067 0.8160997 0.4453241 0.6974236 0.4467378 0.5717941
#>  [50] 0.6705446 0.3706023 0.5105819 0.3762706 0.5893838 0.4819019 0.5542572
#>  [57] 0.4954473 0.5966890 0.3695077 0.4509031 0.4435903 0.6657824 0.4552651
#>  [64] 0.7353777 0.4495056 0.6457412 0.6109975 0.4937252 0.6942978 0.5984635
#>  [71] 0.5005677 0.7143422 0.5277889 0.7172641 0.5291072 0.5275790 0.5653171
#>  [78] 0.5382138 0.6552824 0.5933781 0.5326866 0.5316210 0.4389669 0.6149567
#>  [85] 0.7275354 0.6058395 0.4466954 0.5596024 0.6030269 0.4741519 0.7767047
#>  [92] 0.5830499 0.5640806 0.4412904 0.5082676 0.7150192 0.4172782 0.7887285
#>  [99] 0.7742259 0.6388129 0.6052854 0.4452092 0.6180647 0.4558737 0.5816513
#> [106] 0.3508952 0.5926061 0.5598151 0.5479831 0.7112524 0.5062907 0.4552706
#> [113] 0.4970567 0.4870568 0.5135438 0.5849684 0.6210835 0.6874878 0.5872636
#> [120] 0.6103584 0.7086261 0.6289108 0.6360597 0.5580906 0.5708218 0.6046959
#> [127] 0.6466880 0.6929396 0.5898191 0.5850505 0.6673546 0.4902532 0.3164124
#> [134] 0.7701322 0.4401197 0.5182943 0.5699091 0.4356779 0.5908243 0.6046123
#> [141] 0.5092307 0.6125388 0.5347448 0.6116051 0.3889799 0.5974796 0.6424169
#> [148] 0.5454382 0.4095288 0.6308340 0.5596020 0.6093834 0.6824631 0.6232073
#> [155] 0.6013392 0.4621258 0.6731114 0.5276369 0.7837206 0.4518415 0.4333506
#> [162] 0.6553335 0.4918594 0.5694117 0.5938474 0.6021821 0.5693059 0.5965396
#> [169] 0.4925816 0.6730280 0.7616620 0.4817841 0.6030922 0.5869067 0.6613851
#> [176] 0.7725754 0.7006690 0.5929819 0.4400875 0.5969541 0.4388483 0.6645837
#> [183] 0.6214334 0.7535086 0.5668002 0.6647830 0.4588838 0.5596442 0.6219066
#> [190] 0.4778697 0.5507975 0.5903042 0.4190144 0.4725322 0.6014732 0.7386815
#> [197] 0.4882202 0.5668586 0.4619061 0.5212044 0.7058775 0.5469607 0.6378109
#> [204] 0.7089899 0.4804744 0.5616034 0.5888682 0.6042904 0.5163117 0.5000407
#> [211] 0.6906231 0.5805320 0.6209648 0.4736353 0.5874526 0.6348930 0.5007672
#> [218] 0.4289253 0.5446473 0.5592849 0.4593673 0.5871610 0.4334499 0.5436143
#> [225] 0.5568038 0.6448879 0.4072048 0.4602266 0.6575033 0.5926378 0.6745285
#> [232] 0.5568208 0.6734163 0.4848823 0.6958517 0.6927356 0.7085154 0.4823903
#> [239] 0.5015744 0.5499530 0.7760256 0.5408105 0.7282126 0.6457865 0.4665131
#> [246] 0.6411286 0.5931523 0.4080525 0.5378070 0.6253111 0.6369607 0.6156799
#> [253] 0.6261394 0.6726547 0.5736250 0.5891227 0.5048212 0.5496428 0.4446415
#> [260] 0.5646232 0.5144194 0.4800162 0.7593826 0.5513185 0.6582455 0.4997373
#> [267] 0.5788537 0.4832708 0.5419215 0.6854542 0.4559754 0.5576945 0.5551196
#> [274] 0.6133132 0.5687222 0.5692630 0.4712582 0.6313909 0.7868639 0.5338671
#> [281] 0.4235229 0.7599147 0.5064181 0.4785839 0.6369187 0.6005552 0.4341919
#> [288] 0.5306228 0.5661354 0.6449829 0.5649394 0.5569286 0.4704095 0.5182222
#> [295] 0.4272150 0.5412202 0.3559336 0.3910293 0.4381544 0.6950302 0.7471463
#> [302] 0.3928304 0.5307029 0.5918375 0.4706636 0.5863043 0.6840019 0.5969213
#> [309] 0.5372784 0.5390194 0.5587688 0.5569319 0.4453983 0.5490264 0.4676500
#> [316] 0.6145865 0.8615741 0.7361505 0.5329812 0.6756621 0.5972702 0.7846659
#> [323] 0.6179472 0.5086046 0.4444252 0.5649750 0.5925102 0.5130596 0.3481715
#> [330] 0.6294174 0.6413918 0.5971611 0.6232126 0.5140415 0.4852191 0.5668173
#> [337] 0.6015007 0.5159566 0.4772080 0.9285010 0.6427624 0.5581353 0.7138334
#> [344] 0.4094735 0.5060498 0.4616514 0.3595553 0.4970777 0.7219625 0.4080148
#> [351] 0.5461978 0.4201073 0.4072170 0.5504687 0.6593519 0.5763092 0.6820181
#> [358] 0.4992601 0.6293934 0.3933587 0.5190415 0.5888447 0.3358684 0.6487777
#> [365] 0.7781487 0.6327026 0.6737548 0.3987341 0.5230840 0.5063531 0.7407985
#> [372] 0.5158622 0.4495189 0.3731150 0.4563367 0.5348930 0.5720588 0.6438851
#> [379] 0.4452277 0.5250265 0.7725755 0.6384721 0.6229038 0.6324084 0.5734201
#> [386] 0.7263841 0.5076268 0.4784838 0.5548073 0.5454992 0.4335607 0.6001991
#> [393] 0.5320767 0.7679693 0.6127006 0.5378138 0.5848795 0.5505333 0.6235077
#> [400] 0.4855941 0.5214350 0.4434335 0.5527981 0.4839682 0.5492750 0.5146067
#> [407] 0.4334190 0.3875236 0.4790160 0.7889996 0.6247089 0.3938963 0.5972848
#> [414] 0.6744258 0.5984594 0.5203632 0.4419360 0.6415435 0.4936281 0.5900001
#> [421] 0.5008848 0.6332807 0.5927384 0.6921343 0.6162725 0.5136196 0.4502542
#> [428] 0.4861017 0.6939102 0.5452202 0.4035594 0.5684531 0.5311830 0.4930454
#> [435] 0.5862822 0.7250512 0.6498491 0.6141545 0.3510137 0.5607422 0.3360857
#> [442] 0.5243896 0.5665557 0.3840556 0.6558885 0.5582935 0.6070022 0.4253865
#> [449] 0.5481048 0.6756550 0.5232837 0.3965279 0.6296351 0.6831306 0.4209056
#> [456] 0.5829609 0.4698142 0.5085826 0.3950800 0.5616780 0.6893578 0.5098887
#> [463] 0.6753669 0.6961016 0.7219628 0.5021373 0.5827871 0.5540470 0.5853265
#> [470] 0.6837334 0.6986367 0.4123359 0.5457620 0.6405606 0.6324164 0.6484129
#> [477] 0.4431135 0.4930965 0.5061224 0.6413473 0.5950510 0.6144360 0.7087081
#> [484] 0.4629985 0.5572340 0.5908553 0.5571816 0.7786899 0.4202759 0.4754888
#> [491] 0.6544273 0.5712598 0.5026354 0.5603508 0.5370928 0.4721680 0.5198336
#> [498] 0.4653013 0.5622694 0.6069009
#> 
#> $ci
#>     97.5%      2.5% 
#> 0.2921607 0.6816286 
#>