Uncertainty of KL divergence estimate using Efron's bootstrap.

This function computes a confidence interval for KL divergence based on Efron's bootstrap. The approach only works for kernel density-based estimators since nearest neighbour-based estimators cannot deal with the ties produced when sampling with replacement.

Usage

kld_ci_bootstrap(X, Y, estimator = kld_est_kde1, B = 500L, alpha = 0.05)

Arguments

X, Y

n-by-d and m-by-d matrices, representing n samples from the true distribution \(P\) and m samples from the approximate distribution \(Q\), both in d dimensions. Vector input is treated as a column matrix.

estimator

A function expecting two inputs X and Y, the Kullback-Leibler divergence estimation method. Defaults to kld_est_kde1, which can only deal with one-dimensional two-sample problems (i.e., d = 1 and q = NULL).

B

Number of bootstrap replicates (default: 500), the larger, the more accurate, but also more computationally expensive.

alpha

Error level, defaults to 0.05.

Value

A list with the fields "est" (the estimated KL divergence), "boot" (a length B numeric vector with KL divergence estimates on the bootstrap samples), and "ci" (a length 2 vector containing the lower and upper limits of the estimated confidence interval).

Details

Reference:

Efron, "Bootstrap Methods: Another Look at the Jackknife", The Annals of Statistics, Vol. 7, No. 1 (1979).

Examples

# 1D Gaussian, two samples
X <- rnorm(100)
Y <- rnorm(100, mean = 1, sd = 2)
kld_gaussian(mu1 = 0, sigma1 = 1, mu2 = 1, sigma2 = 2^2)
#> [1] 0.4431472
kld_est_kde1(X, Y)
#> [1] 0.5407894
kld_ci_bootstrap(X, Y)
#> $est
#> [1] 0.5407894
#> 
#> $boot
#>   [1] 0.5668702 0.5899787 0.5398242 0.7825890 0.7282540 0.5836616 0.5323070
#>   [8] 0.5944691 0.5756357 0.7009948 0.5832504 0.6575452 0.4150016 0.4891137
#>  [15] 0.6153804 0.3658000 0.7999569 0.5258166 0.6370202 0.5609504 0.7963517
#>  [22] 0.4855394 0.5625020 0.6821604 0.3070378 0.7902435 0.3425560 0.5990783
#>  [29] 0.8033509 0.4669703 0.5094095 0.6563178 0.3936742 0.6875242 0.5707151
#>  [36] 0.6678754 0.4035469 0.7064625 0.5695642 0.5920236 0.7398587 0.7214493
#>  [43] 0.6189286 0.5335546 0.6024097 0.4463425 0.5972349 0.6037104 0.3759979
#>  [50] 0.5445536 0.4690727 0.3870153 0.6123455 0.6646782 0.3480847 0.4827583
#>  [57] 0.4532060 0.5510635 0.5672788 0.4110953 0.7369949 0.4239394 0.3719737
#>  [64] 0.5435959 0.9177678 0.4391623 0.7260095 0.8597170 0.6276560 0.7503503
#>  [71] 0.4003200 0.3703250 0.5841680 0.7027304 0.7903921 0.5832470 0.8124867
#>  [78] 0.6225012 0.7061191 0.4784415 0.5262107 0.7276147 0.5284607 0.5777535
#>  [85] 0.5176855 0.7329226 0.5833831 0.5142314 0.6674733 0.8069295 0.5154444
#>  [92] 0.4876716 0.5703440 0.4036392 0.6505259 0.5024597 0.8074727 0.6908751
#>  [99] 0.4303372 0.5633785 0.7000042 0.5446193 0.6042019 0.4353443 0.5675316
#> [106] 0.8646448 0.5602715 0.6785901 0.6423015 0.4875504 0.6039583 0.5054726
#> [113] 0.6106811 0.5043729 0.7503555 0.5444447 0.4627628 0.7430295 0.6747777
#> [120] 0.7337021 0.6931405 0.5248339 0.5284347 0.9468242 0.6025044 0.5331212
#> [127] 0.5040548 0.3854860 0.4140398 0.5986617 0.3722265 0.4404812 0.4870566
#> [134] 0.6507716 0.4204857 0.4808526 0.6223634 0.5289857 0.5484617 0.6002559
#> [141] 0.7539531 0.6484273 0.5688443 0.7800582 0.6647371 0.6137652 0.7644880
#> [148] 0.6062097 0.4716108 0.7315988 0.4704843 0.5704661 0.6674782 0.4078620
#> [155] 0.5283076 0.4641248 0.6408602 0.2660723 0.5604488 0.6034856 0.4577510
#> [162] 0.5517004 0.6886191 0.5417409 0.5885914 0.5895763 0.3442747 0.4200235
#> [169] 0.5053560 0.6334796 0.3679558 0.3747509 0.5234716 0.5379211 0.4309694
#> [176] 0.4018167 0.6537812 0.4619953 0.6205340 0.5310369 0.5724315 0.7040679
#> [183] 0.4731337 0.5096845 0.8669339 0.5521965 0.7521446 0.4745745 0.6681081
#> [190] 0.4477185 0.5403012 0.6225969 0.6020522 0.4206679 0.6138097 0.6272720
#> [197] 0.6205215 0.4158588 0.6350209 0.6772048 0.4386508 0.6221336 0.4317992
#> [204] 0.5516424 0.4978053 0.5571593 0.4480694 0.5856695 0.5297308 0.7372416
#> [211] 0.7372641 0.4777188 0.4997775 0.5022414 0.7194156 0.5880705 0.3577307
#> [218] 0.6302598 0.4896798 0.5707569 0.5507919 0.6764853 0.5612685 0.4379802
#> [225] 0.4780651 0.6333513 0.5391444 0.5472408 0.6544738 0.6320120 0.5705789
#> [232] 0.5280460 0.5711666 0.4932459 0.7087491 0.4222780 0.5518719 0.5009012
#> [239] 0.4309541 0.5011525 0.4344990 0.5863332 0.5439101 0.5150105 0.6311514
#> [246] 0.6072637 0.8104840 0.4948440 0.6058435 0.5487532 0.4793671 0.5459669
#> [253] 0.6438157 0.5666778 0.5559063 0.6330704 0.5422704 0.4207944 0.5382654
#> [260] 0.5535454 0.5668519 0.5248431 0.7496024 0.8175819 0.6798159 0.6392881
#> [267] 0.5403900 0.4646443 0.7503195 0.5728163 0.6310828 0.5884900 0.7228281
#> [274] 0.5695739 0.5097918 0.5686416 0.4493077 0.7436161 0.5307757 0.7630660
#> [281] 0.5517998 0.4845222 0.5466180 0.5820217 0.6761519 0.7345479 0.6503394
#> [288] 0.6865946 0.7460274 0.7064958 0.6033085 0.3614747 0.4841070 0.4618232
#> [295] 0.4313030 0.6242694 0.5103870 0.4828880 0.4005591 0.6989522 0.6846306
#> [302] 0.5801245 0.7266238 0.4703705 0.5648281 0.4528533 0.5310459 0.4148869
#> [309] 0.4777567 0.7173456 0.4845906 0.6254410 0.6479978 0.6204672 0.5234715
#> [316] 0.6437954 0.6250128 0.5299742 0.4911031 0.4656587 0.5127976 0.5324828
#> [323] 0.5773116 0.5917022 0.5534020 0.5442377 0.5508337 0.5340183 0.7033968
#> [330] 0.5282172 0.5203106 0.4193979 0.7344173 0.6520042 0.7611250 0.4657213
#> [337] 0.7871883 0.4529739 0.4165320 0.4434836 0.4547977 0.4041774 0.6773584
#> [344] 0.5866676 0.5265922 0.5381415 0.5972894 0.5329037 0.4843615 0.5346369
#> [351] 0.8211379 0.6124778 0.5626477 0.4888535 0.3883668 0.7860406 0.5812876
#> [358] 0.5660244 0.5265930 0.4057618 0.5533372 0.6303250 0.6069675 0.5890758
#> [365] 0.6147003 0.8220991 0.4505745 0.3918250 0.5263879 0.6376915 0.5120046
#> [372] 0.5794125 0.5484909 0.4867031 0.6965119 0.5690835 0.6279927 0.6355766
#> [379] 0.6367680 0.4474490 0.7689258 0.7080755 0.5601888 0.5751444 0.5468522
#> [386] 0.6379364 0.5397343 0.4305261 0.5630522 0.3947505 0.5022550 0.6430958
#> [393] 0.7393579 0.5365093 0.4561416 0.5606133 0.4911586 0.5165187 0.6051956
#> [400] 0.4757832 0.5296890 0.7488677 0.5105777 0.6348452 0.7426182 0.5943910
#> [407] 0.6630648 0.4604238 0.4825927 0.7375628 0.6142615 0.6913835 0.5508594
#> [414] 0.8214474 0.4526122 0.5531261 0.5571333 0.6357829 0.7517061 0.6661748
#> [421] 0.4028231 0.5564133 0.5746179 0.7160126 0.3752662 0.7238221 0.7402956
#> [428] 0.5276447 0.4317261 0.5986535 0.6308539 0.5672529 0.5788352 0.5668367
#> [435] 0.6498697 0.5848032 0.8287415 0.4821081 0.5962786 0.4372410 0.5744726
#> [442] 0.6531774 0.5179686 0.6795868 0.6780413 0.6441993 0.5766338 0.5822416
#> [449] 0.5576204 0.5688284 0.4730181 0.6216361 0.5600381 0.5077849 0.5436644
#> [456] 0.5613140 0.6508331 0.6111794 0.4035919 0.6463858 0.6245598 0.7062704
#> [463] 0.5619097 0.4737595 0.4224011 0.5237036 0.3891060 0.6189676 0.7939376
#> [470] 0.5842204 0.5048788 0.6675332 0.3831204 0.4635420 0.5083612 0.6223514
#> [477] 0.3912910 0.6680349 0.5382379 0.5125338 0.5781548 0.4001934 0.4968399
#> [484] 0.5986025 0.6258779 0.5948365 0.5001101 0.7907376 0.6996749 0.6337403
#> [491] 0.6751333 0.6051370 0.5198627 0.4398362 0.7824523 0.4700644 0.4953939
#> [498] 0.4473259 0.6079068 0.4293870
#> 
#> $ci
#>     97.5%      2.5% 
#> 0.2743641 0.7065831 
#> 
#> $se
#> [1] 0.1129943
#>