Uncertainty of KL divergence estimate using Efron's bootstrap.

This function computes a confidence interval for KL divergence based on Efron's bootstrap. The approach only works for kernel density-based estimators since nearest neighbour-based estimators cannot deal with the ties produced when sampling with replacement.

Usage

kld_ci_bootstrap(X, Y, estimator = kld_est_kde1, B = 500L, alpha = 0.05)

Arguments

X, Y

n-by-d and m-by-d matrices, representing n samples from the true distribution \(P\) and m samples from the approximate distribution \(Q\), both in d dimensions. Vector input is treated as a column matrix.

estimator

A function expecting two inputs X and Y, the Kullback-Leibler divergence estimation method. Defaults to kld_est_kde1, which can only deal with one-dimensional two-sample problems (i.e., d = 1 and q = NULL).

B

Number of bootstrap replicates (default: 500), the larger, the more accurate, but also more computationally expensive.

alpha

Error level, defaults to 0.05.

Value

A list with the fields "est" (the estimated KL divergence), "boot" (a length B numeric vector with KL divergence estimates on the bootstrap samples), and "ci" (a length 2 vector containing the lower and upper limits of the estimated confidence interval).

Details

Reference:

Efron, "Bootstrap Methods: Another Look at the Jackknife", The Annals of Statistics, Vol. 7, No. 1 (1979).

Examples

# 1D Gaussian, two samples
X <- rnorm(100)
Y <- rnorm(100, mean = 1, sd = 2)
kld_gaussian(mu1 = 0, sigma1 = 1, mu2 = 1, sigma2 = 2^2)
#> [1] 0.4431472
kld_est_kde1(X, Y)
#> [1] 0.3934408
kld_ci_bootstrap(X, Y)
#> $est
#> [1] 0.3934408
#> 
#> $boot
#>   [1] 0.4886329 0.4483070 0.4232641 0.3620615 0.5024286 0.3247789 0.5890724
#>   [8] 0.5127156 0.4666501 0.3986273 0.3899138 0.5142621 0.3512228 0.3621353
#>  [15] 0.4609882 0.2431704 0.3333234 0.3809732 0.4037258 0.4812337 0.4149827
#>  [22] 0.5624474 0.3792913 0.5072933 0.3377614 0.5394949 0.4874579 0.4984601
#>  [29] 0.4073645 0.3964788 0.4546963 0.2907430 0.4587043 0.3248458 0.4248938
#>  [36] 0.4088349 0.4136520 0.5285499 0.3577385 0.5321580 0.4729660 0.3073300
#>  [43] 0.3248748 0.3185047 0.3932537 0.3049262 0.3431770 0.4234037 0.3596289
#>  [50] 0.5300989 0.4345493 0.4705799 0.4512874 0.6125461 0.2758610 0.3932345
#>  [57] 0.4464103 0.5343487 0.5033341 0.6389338 0.4499595 0.4588224 0.5868964
#>  [64] 0.4014555 0.5925237 0.3686151 0.4478998 0.4194886 0.5624580 0.4508342
#>  [71] 0.4277825 0.5367627 0.4027787 0.8109231 0.3942527 0.3574392 0.5935286
#>  [78] 0.4875295 0.4325926 0.3539418 0.4305397 0.3459871 0.3469749 0.3921713
#>  [85] 0.3431420 0.5991984 0.3998306 0.6377260 0.3957341 0.5380287 0.4153252
#>  [92] 0.3777003 0.4496862 0.4425852 0.6021612 0.3847954 0.3119121 0.4491915
#>  [99] 0.5339704 0.4213324 0.4273720 0.5589864 0.3782274 0.3223902 0.4789983
#> [106] 0.4847712 0.3863508 0.3233168 0.4296727 0.3697199 0.3791807 0.3751872
#> [113] 0.4515264 0.4714660 0.4144020 0.4871517 0.4127353 0.4905011 0.4572996
#> [120] 0.5431174 0.4388609 0.3886652 0.4840604 0.4051513 0.4889390 0.3508135
#> [127] 0.4526828 0.3604485 0.3586931 0.3468218 0.3037172 0.4331172 0.5825941
#> [134] 0.5724989 0.3722806 0.6710964 0.3443965 0.3759596 0.4556057 0.2991018
#> [141] 0.4075453 0.3780365 0.4768731 0.6456378 0.5722233 0.4723172 0.4574301
#> [148] 0.3445001 0.4302370 0.3933242 0.5288292 0.5480542 0.5392452 0.2731205
#> [155] 0.4696666 0.3699395 0.3886615 0.5092006 0.5117402 0.3655987 0.4581989
#> [162] 0.3868585 0.4006406 0.4613085 0.4691810 0.3996410 0.4067158 0.3756025
#> [169] 0.3820023 0.5115211 0.4797162 0.4296086 0.4228243 0.4267347 0.4449900
#> [176] 0.3516484 0.4934200 0.3290382 0.5169770 0.3703643 0.3210406 0.4059977
#> [183] 0.3247546 0.3498705 0.4008317 0.4201269 0.4191837 0.7159320 0.4585076
#> [190] 0.4790003 0.4939683 0.4510036 0.3720198 0.5129331 0.4497861 0.5082421
#> [197] 0.4665342 0.4423768 0.4468042 0.3858770 0.5129382 0.4139191 0.4905997
#> [204] 0.3739085 0.3850606 0.4717237 0.6702995 0.3287166 0.4131615 0.4847844
#> [211] 0.3917543 0.3614667 0.7112052 0.4289175 0.4739092 0.3518731 0.4033720
#> [218] 0.5844451 0.3430080 0.4797176 0.4209393 0.4998412 0.4484860 0.4545414
#> [225] 0.3843360 0.4565002 0.4719151 0.4833490 0.6912769 0.4424756 0.5141172
#> [232] 0.4103308 0.2399364 0.5429829 0.5048642 0.4485994 0.4763519 0.5700370
#> [239] 0.6926265 0.5496328 0.3523791 0.2640336 0.5297683 0.4143330 0.5075449
#> [246] 0.5329096 0.4611086 0.2763586 0.5073686 0.3962765 0.3335818 0.4615537
#> [253] 0.4589759 0.5397958 0.5018397 0.4279975 0.4295101 0.3159674 0.2980475
#> [260] 0.4856307 0.5707091 0.5232817 0.2414152 0.3253277 0.3806309 0.3627191
#> [267] 0.4855952 0.4882354 0.3001428 0.5818008 0.4144581 0.5342959 0.3930882
#> [274] 0.4289656 0.3980270 0.6231399 0.4224412 0.6268262 0.4918140 0.4770939
#> [281] 0.4536892 0.4202504 0.3816306 0.4041312 0.4149460 0.4145319 0.4252975
#> [288] 0.4664944 0.3611774 0.4561025 0.6304078 0.2652205 0.5685151 0.4111676
#> [295] 0.3138042 0.5160670 0.4605141 0.6217554 0.5311958 0.3986934 0.3954607
#> [302] 0.5399475 0.3103695 0.4442134 0.3228179 0.3293924 0.4143836 0.4767824
#> [309] 0.6541213 0.4210080 0.3208788 0.4185165 0.3754221 0.4647923 0.5264907
#> [316] 0.4993216 0.4035309 0.3101842 0.3967718 0.3194636 0.4068056 0.4704214
#> [323] 0.3711335 0.5869320 0.3305733 0.4119520 0.3617838 0.3474294 0.4545695
#> [330] 0.3831070 0.3337931 0.4242702 0.3765105 0.2693770 0.3424250 0.4171692
#> [337] 0.3469522 0.6332515 0.4285286 0.2876559 0.6198475 0.4308479 0.3773330
#> [344] 0.5179640 0.6297760 0.3436261 0.6015426 0.3646294 0.3341479 0.5814531
#> [351] 0.6407427 0.6025866 0.3284910 0.4373004 0.4804008 0.5205557 0.5062690
#> [358] 0.6052471 0.4265555 0.4741538 0.4708551 0.2183528 0.3319383 0.4775773
#> [365] 0.5314762 0.4031699 0.4217502 0.2970438 0.3923107 0.5519074 0.6674154
#> [372] 0.3817207 0.4274865 0.3809294 0.5825801 0.3661895 0.5645880 0.3683355
#> [379] 0.3215623 0.4360517 0.4594620 0.4669543 0.3385619 0.4785928 0.4825927
#> [386] 0.4208846 0.6402825 0.3139751 0.4135892 0.3988122 0.4211548 0.4094673
#> [393] 0.4297053 0.3395300 0.5050353 0.3885304 0.3118519 0.3899191 0.4647480
#> [400] 0.4137751 0.3972584 0.3368188 0.6251173 0.3070543 0.3306304 0.4520097
#> [407] 0.4308207 0.3811300 0.4510347 0.3265746 0.2963095 0.2058010 0.2816813
#> [414] 0.3069044 0.3004909 0.4120108 0.3287452 0.5453376 0.5439627 0.4757392
#> [421] 0.4805770 0.3199303 0.2936604 0.4217032 0.3396629 0.3085770 0.6239182
#> [428] 0.3856734 0.3037751 0.4370785 0.3014074 0.6272860 0.5245406 0.3988096
#> [435] 0.2435601 0.4907351 0.6328099 0.4014397 0.3499354 0.3187712 0.4412657
#> [442] 0.3631474 0.4116528 0.2739409 0.5060164 0.4281730 0.4166488 0.3732067
#> [449] 0.5106492 0.2470415 0.4154264 0.4241199 0.4464277 0.4385605 0.5102210
#> [456] 0.2879305 0.3564871 0.4205240 0.4437646 0.4496512 0.4101541 0.4134545
#> [463] 0.5667555 0.3384108 0.3251606 0.4067043 0.2294637 0.4444374 0.4765790
#> [470] 0.2911961 0.4286402 0.4184559 0.4005862 0.5145582 0.4188014 0.5647444
#> [477] 0.4497756 0.3776700 0.4518866 0.3263581 0.3983592 0.3159667 0.3785415
#> [484] 0.2865616 0.4038763 0.3820890 0.3389952 0.4083094 0.4719184 0.3121665
#> [491] 0.5535178 0.3029963 0.5711966 0.2536862 0.2494167 0.4007204 0.4367163
#> [498] 0.4701257 0.3603011 0.3693400
#> 
#> $ci
#>     97.5%      2.5% 
#> 0.1485215 0.5157265 
#> 
#> $se
#> [1] 0.09441741
#>