Uncertainty of KL divergence estimate using Efron's bootstrap.

This function computes a confidence interval for KL divergence based on Efron's bootstrap. The approach only works for kernel density-based estimators since nearest neighbour-based estimators cannot deal with the ties produced when sampling with replacement.

Usage

kld_ci_bootstrap(X, Y, estimator = kld_est_kde1, B = 500L, alpha = 0.05)

Arguments

X, Y

n-by-d and m-by-d matrices, representing n samples from the true distribution \(P\) and m samples from the approximate distribution \(Q\), both in d dimensions. Vector input is treated as a column matrix.

estimator

A function expecting two inputs X and Y, the Kullback-Leibler divergence estimation method. Defaults to kld_est_kde1, which can only deal with one-dimensional two-sample problems (i.e., d = 1 and q = NULL).

B

Number of bootstrap replicates (default: 500), the larger, the more accurate, but also more computationally expensive.

alpha

Error level, defaults to 0.05.

Value

A list with the fields "est" (the estimated KL divergence), "boot" (a length B numeric vector with KL divergence estimates on the bootstrap samples), and "ci" (a length 2 vector containing the lower and upper limits of the estimated confidence interval).

Details

Reference:

Efron, "Bootstrap Methods: Another Look at the Jackknife", The Annals of Statistics, Vol. 7, No. 1 (1979).

Examples

# 1D Gaussian, two samples
X <- rnorm(100)
Y <- rnorm(100, mean = 1, sd = 2)
kld_gaussian(mu1 = 0, sigma1 = 1, mu2 = 1, sigma2 = 2^2)
#> [1] 0.4431472
kld_est_kde1(X, Y)
#> [1] 0.3791033
kld_ci_bootstrap(X, Y)
#> $est
#> [1] 0.3791033
#> 
#> $boot
#>   [1] 0.5060680 0.3592553 0.3788807 0.3895601 0.5686834 0.3912618 0.4165747
#>   [8] 0.4156991 0.4184109 0.4211983 0.5147655 0.4560459 0.2923535 0.5610470
#>  [15] 0.4011535 0.4204951 0.4395196 0.3484123 0.4835059 0.3219016 0.5033275
#>  [22] 0.4537540 0.3496277 0.5290075 0.3114643 0.3943844 0.4973195 0.3915794
#>  [29] 0.3099958 0.3910801 0.3535309 0.3648370 0.3957915 0.4010650 0.3599969
#>  [36] 0.3382696 0.2944053 0.4467984 0.4018490 0.4125690 0.5322621 0.3355014
#>  [43] 0.2917067 0.3163182 0.4328579 0.4885382 0.4584021 0.6985732 0.3311069
#>  [50] 0.6012472 0.4129849 0.3458734 0.4561772 0.5042241 0.3346943 0.5379910
#>  [57] 0.2405344 0.4314732 0.3471461 0.4464980 0.4453814 0.3803608 0.3036103
#>  [64] 0.2346441 0.4416357 0.3341583 0.3205120 0.3487911 0.3520064 0.3618340
#>  [71] 0.4449855 0.5336197 0.3856083 0.5185518 0.3648097 0.2863900 0.3611754
#>  [78] 0.5086304 0.3746059 0.3619653 0.3852158 0.5299101 0.2728872 0.3332784
#>  [85] 0.4035171 0.2513046 0.4517366 0.5677504 0.4530967 0.5526383 0.5123424
#>  [92] 0.3021943 0.3222269 0.3373485 0.2978963 0.3873833 0.5516632 0.5536066
#>  [99] 0.4551016 0.3982381 0.2897276 0.5199056 0.3449936 0.3658607 0.3223468
#> [106] 0.3736659 0.3994572 0.4275627 0.3137106 0.4282616 0.4531168 0.2747338
#> [113] 0.3992332 0.5530073 0.4000408 0.3461744 0.3153912 0.4770003 0.2590209
#> [120] 0.4343036 0.5174198 0.3198229 0.2460182 0.4581209 0.2993521 0.5097157
#> [127] 0.4844610 0.4631384 0.5124236 0.6340803 0.4018550 0.2902932 0.4280462
#> [134] 0.6152166 0.3966701 0.3793719 0.3822278 0.3449775 0.4963790 0.3246793
#> [141] 0.4463989 0.4234274 0.5178542 0.4643093 0.3867586 0.5440943 0.2863590
#> [148] 0.4677065 0.3552903 0.3359423 0.4196872 0.4528569 0.4271340 0.6077648
#> [155] 0.5380388 0.4088673 0.3274988 0.4514663 0.3863051 0.4065918 0.3751445
#> [162] 0.4061782 0.4067250 0.2781887 0.2891607 0.3945214 0.5778720 0.4819475
#> [169] 0.4982036 0.5270806 0.5540480 0.3572721 0.3020169 0.5393260 0.3739013
#> [176] 0.4050029 0.3439899 0.4653192 0.5267463 0.4953196 0.4443140 0.3153219
#> [183] 0.4037560 0.3996594 0.3768097 0.2613784 0.3858085 0.3291854 0.3948871
#> [190] 0.4001886 0.3479999 0.4395885 0.5850830 0.4226801 0.3644174 0.4743590
#> [197] 0.5488124 0.2870582 0.2947869 0.3973934 0.5072957 0.3164088 0.7341948
#> [204] 0.4470659 0.3520857 0.4546270 0.2589167 0.3231022 0.1974769 0.5164012
#> [211] 0.3285715 0.3586879 0.3164546 0.4917372 0.3356996 0.4278611 0.4632654
#> [218] 0.5107901 0.3763245 0.4656260 0.3674734 0.4392479 0.3933808 0.4441893
#> [225] 0.4199006 0.5738827 0.4381766 0.3478143 0.3380459 0.2997324 0.2500892
#> [232] 0.4082583 0.3711235 0.3334669 0.2766102 0.4441805 0.4140142 0.2693607
#> [239] 0.5402828 0.3630536 0.3208372 0.5280066 0.2825821 0.4390251 0.4228528
#> [246] 0.3606491 0.4655516 0.5063691 0.3919932 0.4213541 0.4164913 0.5296016
#> [253] 0.4338314 0.5878126 0.6248431 0.3996677 0.4415894 0.5421230 0.2682908
#> [260] 0.3090686 0.4893528 0.3627640 0.4030292 0.5627713 0.3809428 0.3899007
#> [267] 0.4520793 0.7672843 0.4486329 0.4268895 0.4688938 0.4273345 0.4560938
#> [274] 0.5404133 0.4842226 0.3735121 0.3303667 0.3851554 0.3158457 0.4264886
#> [281] 0.3861776 0.4894595 0.3719229 0.4574712 0.4987879 0.5540479 0.3026744
#> [288] 0.5102960 0.3519650 0.5352954 0.3468667 0.3624001 0.4265423 0.4948367
#> [295] 0.4426639 0.4318020 0.4100731 0.3035105 0.3319305 0.3169077 0.4457597
#> [302] 0.4893575 0.3494909 0.4858125 0.4767465 0.5518787 0.3511816 0.4192355
#> [309] 0.3410706 0.3599394 0.3911657 0.4597200 0.5018023 0.3878608 0.2667048
#> [316] 0.4024474 0.4512126 0.4425140 0.4786447 0.2749187 0.5718645 0.4366868
#> [323] 0.3630032 0.3243263 0.4617371 0.4164202 0.4529374 0.3956937 0.3840234
#> [330] 0.2250918 0.4785311 0.3943423 0.4369047 0.3832801 0.4297790 0.3257723
#> [337] 0.4013776 0.3039717 0.4465032 0.3102126 0.4687282 0.3592113 0.5635191
#> [344] 0.2883595 0.4146681 0.5353053 0.3852612 0.4481122 0.3654820 0.3634313
#> [351] 0.3591503 0.2970784 0.6233077 0.5662353 0.2352708 0.3991276 0.4875068
#> [358] 0.5901749 0.4311555 0.4045074 0.3400652 0.4192092 0.4844490 0.4626257
#> [365] 0.4007285 0.5806132 0.3509813 0.4124620 0.4757417 0.4346992 0.6757628
#> [372] 0.3404724 0.3887860 0.4564366 0.6138122 0.4032174 0.4622400 0.5020301
#> [379] 0.2939165 0.4150349 0.2979363 0.6297061 0.3925707 0.4496997 0.3680232
#> [386] 0.4838023 0.3292414 0.4568737 0.3070538 0.4477645 0.5978220 0.2290516
#> [393] 0.2828254 0.4271702 0.3889670 0.5162677 0.4679415 0.3772808 0.3535018
#> [400] 0.3994449 0.3398305 0.3684976 0.3644334 0.4536294 0.4113601 0.5566794
#> [407] 0.3362904 0.5509902 0.6354077 0.3479355 0.3455710 0.4585233 0.3754328
#> [414] 0.4639208 0.3245151 0.4571868 0.5712259 0.5090038 0.3078260 0.4246364
#> [421] 0.4825315 0.4608155 0.4515429 0.4127932 0.4238832 0.4935994 0.4682381
#> [428] 0.3262372 0.3170216 0.5841481 0.4124721 0.3771218 0.3693960 0.3593402
#> [435] 0.4941665 0.2966438 0.4002823 0.3611427 0.3215003 0.5487813 0.4228256
#> [442] 0.4877893 0.4298332 0.3604756 0.3199054 0.3955809 0.4366144 0.3960948
#> [449] 0.4345915 0.4675469 0.5530325 0.5151241 0.5368402 0.3688296 0.6583483
#> [456] 0.4596051 0.3148708 0.4170082 0.4583655 0.3556703 0.5465161 0.3175326
#> [463] 0.4181564 0.3411166 0.3792849 0.4101959 0.4644458 0.4459723 0.3556747
#> [470] 0.4094010 0.4151160 0.3969553 0.3940527 0.4444670 0.3313391 0.6494921
#> [477] 0.3717014 0.5515117 0.4783745 0.4964218 0.3261996 0.3436425 0.4867807
#> [484] 0.3456440 0.2787862 0.3000979 0.4412230 0.3692230 0.3178508 0.4792630
#> [491] 0.4553160 0.5408452 0.4631215 0.3144334 0.3615180 0.4327527 0.3305170
#> [498] 0.3992053 0.4555085 0.4120912
#> 
#> $ci
#>     97.5%      2.5% 
#> 0.1472669 0.4907484 
#>