Uncertainty of KL divergence estimate using Efron's bootstrap.

This function computes a confidence interval for KL divergence based on Efron's bootstrap. The approach only works for kernel density-based estimators since nearest neighbour-based estimators cannot deal with the ties produced when sampling with replacement.

Usage

kld_ci_bootstrap(X, Y, estimator = kld_est_kde1, B = 500L, alpha = 0.05)

Arguments

X, Y

n-by-d and m-by-d matrices, representing n samples from the true distribution \(P\) and m samples from the approximate distribution \(Q\), both in d dimensions. Vector input is treated as a column matrix.

estimator

A function expecting two inputs X and Y, the Kullback-Leibler divergence estimation method. Defaults to kld_est_kde1, which can only deal with one-dimensional two-sample problems (i.e., d = 1 and q = NULL).

B

Number of bootstrap replicates (default: 500), the larger, the more accurate, but also more computationally expensive.

alpha

Error level, defaults to 0.05.

Value

A list with the fields "est" (the estimated KL divergence), "boot" (a length B numeric vector with KL divergence estimates on the bootstrap samples), and "ci" (a length 2 vector containing the lower and upper limits of the estimated confidence interval).

Details

Reference:

Efron, "Bootstrap Methods: Another Look at the Jackknife", The Annals of Statistics, Vol. 7, No. 1 (1979).

Examples

# 1D Gaussian, two samples
X <- rnorm(100)
Y <- rnorm(100, mean = 1, sd = 2)
kld_gaussian(mu1 = 0, sigma1 = 1, mu2 = 1, sigma2 = 2^2)
#> [1] 0.4431472
kld_est_kde1(X, Y)
#> [1] 0.3050227
kld_ci_bootstrap(X, Y)
#> $est
#> [1] 0.3050227
#> 
#> $boot
#>   [1] 0.3405813 0.2759595 0.2989722 0.3608795 0.2820310 0.4089895 0.2037100
#>   [8] 0.2298515 0.2975632 0.2933312 0.3037907 0.4236977 0.2930007 0.4814454
#>  [15] 0.3628959 0.3070997 0.3463923 0.3013489 0.3195314 0.3677802 0.2862936
#>  [22] 0.2585601 0.3719164 0.5271828 0.1952821 0.2785403 0.2547143 0.2901303
#>  [29] 0.3223036 0.3227624 0.3896755 0.2233591 0.2981194 0.3547075 0.3633035
#>  [36] 0.4723061 0.2666349 0.4520678 0.3832863 0.3394399 0.3515564 0.2681370
#>  [43] 0.3186835 0.4303775 0.3395378 0.3108111 0.2894643 0.3749272 0.3754439
#>  [50] 0.3645719 0.2659062 0.3859169 0.3706762 0.4464463 0.3972119 0.4046737
#>  [57] 0.3365859 0.3151764 0.3661885 0.3258413 0.4587875 0.4483034 0.3100639
#>  [64] 0.3723094 0.3222541 0.3568858 0.2405015 0.5120247 0.2423917 0.2869291
#>  [71] 0.2970331 0.4579056 0.3059216 0.3044820 0.2898212 0.3044169 0.3076777
#>  [78] 0.1431217 0.2779912 0.2933476 0.3848355 0.2095176 0.2946207 0.2678599
#>  [85] 0.4036329 0.4389168 0.3484479 0.4804301 0.3803210 0.3199658 0.2862204
#>  [92] 0.2165853 0.3665042 0.2116521 0.3005231 0.2690742 0.1761013 0.2363471
#>  [99] 0.2828429 0.3772721 0.3324527 0.4614082 0.2698770 0.4321194 0.3250169
#> [106] 0.2762328 0.2767493 0.3945521 0.3867418 0.3398538 0.3735820 0.2410586
#> [113] 0.2542410 0.4217269 0.3274935 0.2373387 0.3945837 0.4477456 0.4944111
#> [120] 0.3812752 0.3859898 0.3197029 0.3046155 0.3361809 0.2178093 0.3988175
#> [127] 0.2573824 0.3355638 0.2753404 0.5304920 0.3531091 0.3904193 0.3835461
#> [134] 0.2487991 0.2820739 0.2971148 0.4035106 0.5254403 0.3157678 0.2522565
#> [141] 0.4543214 0.2904147 0.3446668 0.4080520 0.3359239 0.3356632 0.4277504
#> [148] 0.2883702 0.3643804 0.2990373 0.3011253 0.4408351 0.3980504 0.3050627
#> [155] 0.3525360 0.2541841 0.2180605 0.5084290 0.5099041 0.3690669 0.2907935
#> [162] 0.3163497 0.3533737 0.4051287 0.3843182 0.2922758 0.3929340 0.4031213
#> [169] 0.2825171 0.4130931 0.3135409 0.1750810 0.2991844 0.3422507 0.3111762
#> [176] 0.2837979 0.2092655 0.3606812 0.2528800 0.2655454 0.3115404 0.1632055
#> [183] 0.2787814 0.2121136 0.4191858 0.4445027 0.3576559 0.1880795 0.2932833
#> [190] 0.1910691 0.4046296 0.2980291 0.3264257 0.2728164 0.2610559 0.3645869
#> [197] 0.4371691 0.3427853 0.3842085 0.2963402 0.4033704 0.2805200 0.3649271
#> [204] 0.3889403 0.2660165 0.2783413 0.4090770 0.3418836 0.3588062 0.4130792
#> [211] 0.4662642 0.2564986 0.2699100 0.3206022 0.4270716 0.2805487 0.3828443
#> [218] 0.3031060 0.2406037 0.4913461 0.4219166 0.2478065 0.2773810 0.4969978
#> [225] 0.3464451 0.3876750 0.3567274 0.2652825 0.4724944 0.4173217 0.3647149
#> [232] 0.3131505 0.3515452 0.1954278 0.3063946 0.3134114 0.1338511 0.5185290
#> [239] 0.4097246 0.3330705 0.3844697 0.3671858 0.4772020 0.5231793 0.4161311
#> [246] 0.3849208 0.4180833 0.3760193 0.4740598 0.4157970 0.4045764 0.3701897
#> [253] 0.2959801 0.2377789 0.3626900 0.3901289 0.3296153 0.3654748 0.1861737
#> [260] 0.2386697 0.3698382 0.2662481 0.2478252 0.2300907 0.2824883 0.2803773
#> [267] 0.2619032 0.5547200 0.2853126 0.3238336 0.4491149 0.4174727 0.3793103
#> [274] 0.2525691 0.2616589 0.2869906 0.3948174 0.1793082 0.4274577 0.5426098
#> [281] 0.2649774 0.3919666 0.1941403 0.2106423 0.3219387 0.2921685 0.3203284
#> [288] 0.3163798 0.3772839 0.3549205 0.4691387 0.2081722 0.2849600 0.4769738
#> [295] 0.3826385 0.2823862 0.4889773 0.3385636 0.2982156 0.3273983 0.3181520
#> [302] 0.5160549 0.3132587 0.3497198 0.3497052 0.3790155 0.1331927 0.2457246
#> [309] 0.1882233 0.3638202 0.2427906 0.4237345 0.3465594 0.4187545 0.3532372
#> [316] 0.2393409 0.2090690 0.3732129 0.2159273 0.2547720 0.3572695 0.1613210
#> [323] 0.5163393 0.1799785 0.2176596 0.3560761 0.2709166 0.4147612 0.3577543
#> [330] 0.2902213 0.3907957 0.3838585 0.3545619 0.3216806 0.2016666 0.2762711
#> [337] 0.2099635 0.3600993 0.1575679 0.4219726 0.2690810 0.2921313 0.2357281
#> [344] 0.2686747 0.4238348 0.5457970 0.2617537 0.3875186 0.3872805 0.2614699
#> [351] 0.2296432 0.4669377 0.2755127 0.3830154 0.4766764 0.2993211 0.3187374
#> [358] 0.4208868 0.4482549 0.3996159 0.2222572 0.3098973 0.3231171 0.4322778
#> [365] 0.2410444 0.2611379 0.2722254 0.2818188 0.2267637 0.2874799 0.3900068
#> [372] 0.3310021 0.3883961 0.2854581 0.3442762 0.2879518 0.3465131 0.3461138
#> [379] 0.3346817 0.3919499 0.3254677 0.2877063 0.3768667 0.3287753 0.2588276
#> [386] 0.3060675 0.3604153 0.3032699 0.4830839 0.2657079 0.3496215 0.2665833
#> [393] 0.4547882 0.3734938 0.2863747 0.4017252 0.3517413 0.2902519 0.4682025
#> [400] 0.3932413 0.2595632 0.3364470 0.2527098 0.4159228 0.2880143 0.4754935
#> [407] 0.2174823 0.3770082 0.3263284 0.3662489 0.3826770 0.3606107 0.4114491
#> [414] 0.3662332 0.3109566 0.3347341 0.3446104 0.2364224 0.3902022 0.3073725
#> [421] 0.4533657 0.4691594 0.2080741 0.3178942 0.3601084 0.4320870 0.3058481
#> [428] 0.2737960 0.2850038 0.2721114 0.2130004 0.3217862 0.2426364 0.2108803
#> [435] 0.3818730 0.4334415 0.3196980 0.4385764 0.3889767 0.5194171 0.2975113
#> [442] 0.2726280 0.3525890 0.2379464 0.3889349 0.3231751 0.2690715 0.2437604
#> [449] 0.4038966 0.2023297 0.5137679 0.4309827 0.3678088 0.5160105 0.2480808
#> [456] 0.3189772 0.2758643 0.3493316 0.4547622 0.5029576 0.3690102 0.3696709
#> [463] 0.2929274 0.4632789 0.4211752 0.3196247 0.4599272 0.5721914 0.2099800
#> [470] 0.2722660 0.4375804 0.2845087 0.3686556 0.2406192 0.3287835 0.2926240
#> [477] 0.3099540 0.3373231 0.2379550 0.4623107 0.2923541 0.2955730 0.2841066
#> [484] 0.3903094 0.2802879 0.3575215 0.2411001 0.4835875 0.4646308 0.3827921
#> [491] 0.3631770 0.4366889 0.4240264 0.3550525 0.2435634 0.3241397 0.3406431
#> [498] 0.1984721 0.2582185 0.4124178
#> 
#> $ci
#>     97.5%      2.5% 
#> 0.0951002 0.4204704 
#>