Uncertainty of KL divergence estimate using Efron's bootstrap.

This function computes a confidence interval for KL divergence based on Efron's bootstrap. The approach only works for kernel density-based estimators since nearest neighbour-based estimators cannot deal with the ties produced when sampling with replacement.

Usage

kld_ci_bootstrap(X, Y, estimator = kld_est_kde1, B = 500L, alpha = 0.05)

Arguments

X, Y

n-by-d and m-by-d matrices, representing n samples from the true distribution \(P\) and m samples from the approximate distribution \(Q\), both in d dimensions. Vector input is treated as a column matrix.

estimator

A function expecting two inputs X and Y, the Kullback-Leibler divergence estimation method. Defaults to kld_est_kde1, which can only deal with one-dimensional two-sample problems (i.e., d = 1 and q = NULL).

B

Number of bootstrap replicates (default: 500), the larger, the more accurate, but also more computationally expensive.

alpha

Error level, defaults to 0.05.

Value

A list with the fields "est" (the estimated KL divergence), "boot" (a length B numeric vector with KL divergence estimates on the bootstrap samples), and "ci" (a length 2 vector containing the lower and upper limits of the estimated confidence interval).

Details

Reference:

Efron, "Bootstrap Methods: Another Look at the Jackknife", The Annals of Statistics, Vol. 7, No. 1 (1979).

Examples

# 1D Gaussian, two samples
X <- rnorm(100)
Y <- rnorm(100, mean = 1, sd = 2)
kld_gaussian(mu1 = 0, sigma1 = 1, mu2 = 1, sigma2 = 2^2)
#> [1] 0.4431472
kld_est_kde1(X, Y)
#> [1] 0.406161
kld_ci_bootstrap(X, Y)
#> $est
#> [1] 0.406161
#> 
#> $boot
#>   [1] 0.5230979 0.4393730 0.2084822 0.5178321 0.3320123 0.5422744 0.4335514
#>   [8] 0.5265660 0.4844884 0.5203872 0.5479830 0.2910881 0.2874066 0.2417710
#>  [15] 0.4430562 0.3426571 0.5460419 0.3983635 0.3525426 0.6277599 0.3428653
#>  [22] 0.5013628 0.5993057 0.2731118 0.4469808 0.4228061 0.5054455 0.3707455
#>  [29] 0.5454898 0.5295339 0.4410906 0.4174851 0.4674892 0.4640234 0.5221416
#>  [36] 0.4994805 0.5819754 0.7245144 0.5203354 0.2695040 0.3045619 0.3247180
#>  [43] 0.2493399 0.4764955 0.2778604 0.5109933 0.3884312 0.4137779 0.3276550
#>  [50] 0.3670403 0.4434579 0.5119172 0.5142030 0.4589871 0.3565249 0.3424829
#>  [57] 0.4683067 0.4033137 0.4072842 0.4795268 0.5493562 0.2664945 0.5155781
#>  [64] 0.3262669 0.3243632 0.4589371 0.4275151 0.4181995 0.4159453 0.4167093
#>  [71] 0.4479202 0.3329494 0.3993621 0.6087838 0.5924195 0.4507199 0.2837427
#>  [78] 0.5839131 0.3730942 0.4396509 0.3303989 0.3614977 0.4124947 0.3677006
#>  [85] 0.5247840 0.3759459 0.6154283 0.4054579 0.5579409 0.3843039 0.6370976
#>  [92] 0.3640355 0.4026029 0.4812376 0.5830506 0.3600092 0.4021886 0.3754068
#>  [99] 0.4750704 0.4333947 0.4636797 0.3635437 0.5166840 0.4313571 0.3158741
#> [106] 0.5800207 0.4453621 0.3378079 0.3664467 0.5577741 0.5831629 0.4545442
#> [113] 0.4994597 0.3149995 0.4500764 0.4115278 0.6137863 0.4575573 0.3627522
#> [120] 0.3564707 0.5249216 0.3672608 0.4229275 0.4381762 0.4892712 0.4146551
#> [127] 0.3235685 0.4699887 0.4251512 0.3163590 0.3267922 0.4550693 0.3428165
#> [134] 0.3814306 0.4754864 0.4329351 0.3531284 0.2863068 0.5112751 0.3873269
#> [141] 0.3049648 0.4069854 0.3848640 0.5980387 0.4545435 0.5777789 0.5065278
#> [148] 0.6512245 0.3915898 0.3853358 0.5615733 0.4773894 0.3813402 0.3431646
#> [155] 0.4892596 0.4603514 0.4151518 0.4575436 0.4539138 0.5399722 0.4716611
#> [162] 0.4430217 0.6858786 0.4206351 0.2814670 0.3446271 0.5314682 0.3755677
#> [169] 0.4263336 0.4934682 0.6433130 0.4110577 0.3271813 0.4557481 0.5618537
#> [176] 0.6430892 0.4760751 0.4672846 0.6939852 0.3752904 0.4292151 0.5129862
#> [183] 0.3767386 0.3706822 0.5251042 0.3021578 0.3794353 0.5845238 0.4663295
#> [190] 0.6293486 0.3854216 0.4451768 0.5061413 0.3598774 0.4514920 0.5723757
#> [197] 0.5079489 0.3524508 0.6126566 0.5951558 0.6261836 0.4256995 0.2871832
#> [204] 0.5953690 0.4609299 0.3217768 0.3602135 0.4240764 0.3212919 0.5109236
#> [211] 0.4111118 0.4011760 0.4623192 0.4241068 0.4116908 0.6070502 0.5180188
#> [218] 0.3679994 0.4887688 0.5533697 0.2477289 0.5245508 0.4907683 0.3739798
#> [225] 0.3703305 0.3917708 0.5982495 0.2963264 0.3219281 0.3866264 0.5081139
#> [232] 0.4701577 0.4308317 0.2922698 0.4443218 0.4529126 0.4752298 0.5815876
#> [239] 0.5309420 0.4314631 0.5398479 0.5235205 0.2880330 0.3380807 0.3415867
#> [246] 0.4843590 0.6039315 0.5184893 0.3829524 0.4383750 0.5943567 0.5217949
#> [253] 0.4755660 0.3371067 0.4270879 0.3759193 0.4225169 0.2796217 0.5067293
#> [260] 0.1789483 0.4689848 0.3358639 0.5370640 0.4859863 0.5217873 0.3288797
#> [267] 0.4018819 0.4388574 0.5660262 0.4279246 0.5101683 0.4523491 0.2962077
#> [274] 0.4606826 0.5138737 0.5029940 0.5023342 0.3980264 0.4947316 0.4941713
#> [281] 0.4729334 0.5289379 0.5363433 0.4623799 0.3757886 0.4149585 0.4435031
#> [288] 0.4289077 0.6015538 0.2733844 0.4332792 0.5312016 0.6505465 0.4356618
#> [295] 0.4786240 0.4988193 0.4024459 0.4647321 0.2673552 0.4013178 0.5205750
#> [302] 0.3617355 0.4225155 0.4846519 0.3078108 0.3065830 0.5565103 0.3278611
#> [309] 0.5980828 0.3651416 0.4561186 0.4577948 0.4467831 0.4067339 0.2599752
#> [316] 0.5115255 0.3333856 0.3628953 0.4128024 0.5303891 0.4065713 0.4114578
#> [323] 0.5099714 0.4121307 0.4441802 0.4266588 0.4182770 0.3258964 0.6496202
#> [330] 0.3645628 0.3479574 0.3062401 0.5137752 0.6562248 0.4296550 0.4027241
#> [337] 0.4663947 0.3687082 0.4967165 0.4793792 0.3128425 0.5602185 0.2420246
#> [344] 0.3584381 0.5199746 0.3746610 0.3762874 0.4211199 0.3313830 0.3267393
#> [351] 0.4498939 0.3720484 0.4842561 0.5106005 0.3318381 0.3738934 0.5883835
#> [358] 0.3788955 0.3978794 0.3289527 0.4147763 0.4368779 0.3979281 0.3620641
#> [365] 0.4680235 0.2172004 0.3585065 0.4221041 0.3531908 0.4459576 0.5606075
#> [372] 0.3834731 0.4276247 0.4779198 0.3679894 0.4789207 0.4501043 0.3742948
#> [379] 0.4276048 0.5150501 0.3846689 0.3373926 0.4162075 0.4714897 0.4822132
#> [386] 0.5885624 0.4232741 0.3256588 0.5135632 0.3415176 0.5241561 0.5102055
#> [393] 0.5045825 0.3894827 0.3948634 0.3573106 0.4640221 0.5401484 0.4008881
#> [400] 0.4423494 0.3464449 0.2895972 0.6228074 0.3732997 0.3179045 0.3456234
#> [407] 0.2986777 0.3068138 0.3315152 0.4786886 0.3391016 0.4967354 0.4966359
#> [414] 0.4838702 0.4824179 0.3553499 0.4590674 0.3676031 0.3631635 0.4058323
#> [421] 0.4032137 0.2605882 0.4249857 0.3542286 0.4981752 0.7556413 0.4230778
#> [428] 0.3917260 0.4216928 0.4244228 0.2913993 0.5988797 0.3347671 0.4515446
#> [435] 0.5616637 0.4803447 0.3687168 0.2777824 0.3825558 0.4061556 0.4468472
#> [442] 0.3981121 0.4904180 0.4476605 0.3917741 0.5448148 0.3858763 0.4710426
#> [449] 0.2534708 0.3242880 0.4678259 0.4418223 0.4664635 0.5375707 0.4597706
#> [456] 0.5224707 0.4534189 0.4303384 0.3046450 0.3941963 0.4961247 0.5703297
#> [463] 0.3686549 0.2090944 0.2954026 0.4949832 0.5536525 0.5168620 0.5077207
#> [470] 0.4074866 0.3293512 0.3492322 0.4564076 0.5012213 0.2570590 0.4484659
#> [477] 0.4157523 0.3851749 0.4506556 0.4242897 0.2441827 0.6460897 0.3949120
#> [484] 0.4920496 0.3691786 0.3922914 0.4823163 0.4011708 0.2855178 0.3549478
#> [491] 0.4189331 0.3752599 0.4062079 0.3786223 0.5237807 0.4319362 0.5379058
#> [498] 0.3297882 0.6451029 0.5781968
#> 
#> $ci
#>     97.5%      2.5% 
#> 0.1789052 0.5489284 
#> 
#> $se
#> [1] 0.0951857
#>