Uncertainty of KL divergence estimate using Efron's bootstrap.

This function computes a confidence interval for KL divergence based on Efron's bootstrap. The approach only works for kernel density-based estimators since nearest neighbour-based estimators cannot deal with the ties produced when sampling with replacement.

Usage

kld_ci_bootstrap(X, Y, estimator = kld_est_kde1, B = 500L, alpha = 0.05)

Arguments

X, Y

n-by-d and m-by-d matrices, representing n samples from the true distribution \(P\) and m samples from the approximate distribution \(Q\), both in d dimensions. Vector input is treated as a column matrix.

estimator

A function expecting two inputs X and Y, the Kullback-Leibler divergence estimation method. Defaults to kld_est_kde1, which can only deal with one-dimensional two-sample problems (i.e., d = 1 and q = NULL).

B

Number of bootstrap replicates (default: 500), the larger, the more accurate, but also more computationally expensive.

alpha

Error level, defaults to 0.05.

Value

A list with the fields "est" (the estimated KL divergence), "boot" (a length B numeric vector with KL divergence estimates on the bootstrap samples), and "ci" (a length 2 vector containing the lower and upper limits of the estimated confidence interval).

Details

Reference:

Efron, "Bootstrap Methods: Another Look at the Jackknife", The Annals of Statistics, Vol. 7, No. 1 (1979).

Examples

# 1D Gaussian, two samples
X <- rnorm(100)
Y <- rnorm(100, mean = 1, sd = 2)
kld_gaussian(mu1 = 0, sigma1 = 1, mu2 = 1, sigma2 = 2^2)
#> [1] 0.4431472
kld_est_kde1(X, Y)
#> [1] 0.38337
kld_ci_bootstrap(X, Y)
#> $est
#> [1] 0.38337
#> 
#> $boot
#>   [1] 0.3864057 0.4437857 0.3187424 0.3890996 0.5574449 0.4015694 0.2637385
#>   [8] 0.4125962 0.4004931 0.3774251 0.3044314 0.4955670 0.3876667 0.2730488
#>  [15] 0.4186253 0.4584970 0.3952531 0.4471842 0.2595159 0.5697790 0.3478710
#>  [22] 0.4975457 0.4509654 0.2577853 0.3445707 0.3747971 0.4379092 0.5020689
#>  [29] 0.3849890 0.4736266 0.4705213 0.4929082 0.3271171 0.3386747 0.3567620
#>  [36] 0.4682793 0.5150908 0.4698681 0.4148704 0.4886892 0.3660747 0.3240403
#>  [43] 0.4596540 0.4106522 0.4950784 0.4898927 0.5182723 0.3138686 0.5253524
#>  [50] 0.4910370 0.3132865 0.3318893 0.3501046 0.3731898 0.3696680 0.4102511
#>  [57] 0.4994622 0.3731863 0.3880325 0.3570120 0.4367755 0.3970752 0.3245815
#>  [64] 0.2917556 0.4253449 0.2882896 0.4963045 0.4778063 0.4282561 0.3156825
#>  [71] 0.4157658 0.4664087 0.3333814 0.3845334 0.5443481 0.4927093 0.4589624
#>  [78] 0.2823737 0.4355877 0.2801104 0.2619558 0.4656349 0.5253154 0.4009918
#>  [85] 0.3912298 0.3773911 0.3167195 0.3333787 0.3793243 0.4121692 0.5852517
#>  [92] 0.2780728 0.4123542 0.4657759 0.5570082 0.3298519 0.3457481 0.3620235
#>  [99] 0.5187590 0.3664122 0.4988896 0.3087960 0.4371262 0.5548134 0.4954401
#> [106] 0.4425785 0.4014021 0.4057419 0.3590797 0.3863063 0.4758426 0.4146160
#> [113] 0.5270696 0.3640590 0.4112185 0.3588679 0.3509737 0.4225153 0.4699437
#> [120] 0.5001352 0.3415917 0.3381389 0.4208039 0.4514730 0.5741548 0.3841608
#> [127] 0.3659861 0.3752826 0.6067927 0.4241120 0.4049611 0.4915640 0.3366347
#> [134] 0.4013166 0.4815084 0.4762599 0.4816306 0.3882633 0.4345823 0.6095030
#> [141] 0.2149796 0.4960927 0.4553929 0.2623129 0.2860333 0.2064714 0.6877273
#> [148] 0.5307221 0.3117437 0.3882630 0.2788024 0.2594049 0.2982000 0.4395885
#> [155] 0.5067298 0.5476301 0.3989697 0.3057951 0.5116443 0.2881853 0.4654539
#> [162] 0.3982801 0.2941671 0.5805359 0.4242004 0.4869666 0.3260245 0.5187440
#> [169] 0.4667752 0.4226753 0.4345728 0.3898801 0.5382334 0.5714428 0.4266235
#> [176] 0.3967310 0.3964435 0.3859078 0.5203237 0.3546468 0.4451969 0.5195449
#> [183] 0.2695533 0.4878936 0.3393727 0.3781968 0.3555755 0.5375520 0.4288761
#> [190] 0.2686903 0.4034760 0.4328801 0.4366650 0.3721303 0.3993530 0.4645305
#> [197] 0.5394104 0.3869312 0.3190528 0.4709041 0.3809036 0.4641894 0.7212246
#> [204] 0.4360090 0.3564876 0.4959269 0.3183052 0.3684200 0.3503511 0.4185180
#> [211] 0.2458156 0.4070819 0.3523376 0.4879825 0.4731050 0.3202993 0.5437477
#> [218] 0.2724390 0.4114818 0.5136147 0.4246861 0.3115384 0.6384103 0.3973632
#> [225] 0.3878589 0.3662092 0.4383957 0.3367717 0.4935221 0.3228161 0.4289223
#> [232] 0.4730213 0.4838732 0.3855814 0.2689555 0.2792412 0.5299603 0.5174934
#> [239] 0.4545190 0.3847836 0.3484924 0.7022780 0.3199647 0.4893216 0.4296942
#> [246] 0.3574060 0.5600362 0.3666587 0.4142801 0.4610201 0.5176512 0.4740274
#> [253] 0.4177124 0.5630779 0.3664554 0.3693964 0.4166811 0.4571911 0.4670549
#> [260] 0.4868589 0.2502768 0.3777426 0.4743346 0.5395670 0.3308748 0.4887513
#> [267] 0.3797411 0.2911900 0.5331270 0.3368592 0.5926046 0.3515599 0.3892877
#> [274] 0.5646418 0.4131515 0.3551471 0.3500462 0.3963354 0.3594921 0.5207426
#> [281] 0.4801956 0.4910917 0.6416329 0.5295753 0.3382192 0.4546326 0.4744324
#> [288] 0.5597415 0.3667436 0.5859734 0.3829190 0.2946546 0.4528785 0.4359283
#> [295] 0.5043457 0.5010870 0.2682772 0.3372381 0.3736518 0.4053176 0.3841810
#> [302] 0.2906701 0.3552224 0.3622820 0.4002046 0.4038321 0.4724343 0.4932204
#> [309] 0.4139003 0.3991205 0.3567206 0.3613621 0.3155873 0.4157838 0.2965440
#> [316] 0.5614014 0.3639910 0.3907839 0.2209848 0.4030467 0.4295454 0.3257005
#> [323] 0.4319685 0.4436591 0.3574394 0.4880606 0.3596015 0.3739288 0.4274981
#> [330] 0.2979286 0.4530652 0.3807332 0.4667085 0.3532656 0.4022974 0.3207531
#> [337] 0.3719558 0.3756276 0.4552123 0.4022563 0.3943901 0.3701890 0.4349382
#> [344] 0.2401864 0.4116908 0.4811161 0.3060635 0.4331589 0.2975264 0.4753523
#> [351] 0.5134855 0.3088706 0.4030039 0.2901912 0.3953015 0.6178286 0.4059073
#> [358] 0.3311704 0.4514532 0.5741624 0.3514138 0.4526362 0.4957559 0.3604575
#> [365] 0.4248426 0.4113077 0.3822183 0.3984681 0.4690874 0.6336653 0.4247030
#> [372] 0.3584574 0.5952328 0.2899239 0.4241265 0.3737185 0.5338357 0.4166586
#> [379] 0.2903482 0.3835395 0.2466335 0.3408519 0.4554365 0.6288604 0.4062125
#> [386] 0.3054324 0.3384048 0.3940045 0.4466839 0.4855499 0.4030018 0.3536696
#> [393] 0.3064321 0.3688391 0.4575399 0.4022880 0.3058364 0.4248331 0.4874022
#> [400] 0.4189426 0.4221072 0.4882658 0.3249944 0.4777010 0.2273724 0.5013123
#> [407] 0.4554984 0.3043266 0.4623057 0.5288672 0.4481957 0.3934807 0.3174466
#> [414] 0.4197058 0.3884966 0.4383923 0.5465630 0.4503598 0.5566545 0.4996967
#> [421] 0.3716988 0.3364906 0.4802170 0.4017481 0.3349678 0.5041127 0.4764474
#> [428] 0.5562453 0.3804797 0.5217439 0.4580192 0.4763687 0.3677796 0.4436042
#> [435] 0.4496900 0.4652037 0.4739816 0.3289532 0.4705658 0.3614492 0.4424760
#> [442] 0.2807957 0.4525874 0.3826192 0.5067752 0.2937666 0.3554545 0.3560572
#> [449] 0.2993814 0.4784646 0.3948511 0.4949649 0.3212696 0.4408759 0.4193176
#> [456] 0.5271549 0.5736278 0.5478475 0.3518080 0.3379571 0.5127873 0.4670331
#> [463] 0.5081119 0.3131053 0.3873697 0.3584938 0.5703624 0.4358197 0.2778686
#> [470] 0.3676641 0.4303294 0.4073072 0.3278213 0.4304904 0.5447833 0.3971325
#> [477] 0.6975783 0.3407513 0.4095137 0.3514560 0.4653776 0.4973907 0.2142661
#> [484] 0.5494517 0.2338247 0.3752271 0.3682427 0.4257369 0.4367548 0.3591730
#> [491] 0.3417286 0.3685936 0.3277360 0.4669297 0.3109208 0.3553415 0.4192011
#> [498] 0.5199040 0.3417890 0.3827863
#> 
#> $ci
#>     97.5%      2.5% 
#> 0.1772853 0.5060652 
#>