Uncertainty of KL divergence estimate using Efron's bootstrap.

This function computes a confidence interval for KL divergence based on Efron's bootstrap. The approach only works for kernel density-based estimators since nearest neighbour-based estimators cannot deal with the ties produced when sampling with replacement.

Usage

kld_ci_bootstrap(X, Y, estimator = kld_est_kde1, B = 500L, alpha = 0.05)

Arguments

X, Y

n-by-d and m-by-d matrices, representing n samples from the true distribution \(P\) and m samples from the approximate distribution \(Q\), both in d dimensions. Vector input is treated as a column matrix.

estimator

A function expecting two inputs X and Y, the Kullback-Leibler divergence estimation method. Defaults to kld_est_kde1, which can only deal with one-dimensional two-sample problems (i.e., d = 1 and q = NULL).

B

Number of bootstrap replicates (default: 500), the larger, the more accurate, but also more computationally expensive.

alpha

Error level, defaults to 0.05.

Value

A list with the fields "est" (the estimated KL divergence), "boot" (a length B numeric vector with KL divergence estimates on the bootstrap samples), and "ci" (a length 2 vector containing the lower and upper limits of the estimated confidence interval).

Details

Reference:

Efron, "Bootstrap Methods: Another Look at the Jackknife", The Annals of Statistics, Vol. 7, No. 1 (1979).

Examples

# 1D Gaussian, two samples
X <- rnorm(100)
Y <- rnorm(100, mean = 1, sd = 2)
kld_gaussian(mu1 = 0, sigma1 = 1, mu2 = 1, sigma2 = 2^2)
#> [1] 0.4431472
kld_est_kde1(X, Y)
#> [1] 0.4742318
kld_ci_bootstrap(X, Y)
#> $est
#> [1] 0.4742318
#> 
#> $boot
#>   [1] 0.7268792 0.5701810 0.5142156 0.6963252 0.6168621 0.4534190 0.5319263
#>   [8] 0.4264821 0.4965600 0.5749274 0.4914559 0.4122523 0.4750368 0.5908466
#>  [15] 0.5896513 0.4535925 0.4123061 0.5467812 0.6252856 0.6587475 0.5592077
#>  [22] 0.4706688 0.2847413 0.4706661 0.4487507 0.4041206 0.4949388 0.5564426
#>  [29] 0.4238606 0.3980932 0.5380378 0.5374383 0.6772957 0.6110023 0.5029197
#>  [36] 0.6421515 0.5909363 0.5812507 0.6213512 0.5244721 0.4464640 0.4105032
#>  [43] 0.6129854 0.5988591 0.6353158 0.4846607 0.6152768 0.4550395 0.6114732
#>  [50] 0.4725927 0.4609120 0.6091877 0.7168709 0.4758790 0.7657868 0.5101467
#>  [57] 0.6075965 0.3212671 0.4809555 0.4590775 0.3529370 0.5860860 0.5384212
#>  [64] 0.6515393 0.4894546 0.4333627 0.5883312 0.4648523 0.4843993 0.7510941
#>  [71] 0.4955555 0.3581731 0.5486965 0.5831307 0.5691179 0.7262431 0.5355411
#>  [78] 0.6263840 0.4041037 0.4743077 0.7008684 0.6214790 0.5704322 0.4012176
#>  [85] 0.5613289 0.6229310 0.5933526 0.3241398 0.5379907 0.4643824 0.4714580
#>  [92] 0.4416337 0.4911904 0.5588434 0.4148489 0.3661578 0.3915477 0.5350162
#>  [99] 0.4617430 0.4494433 0.5175862 0.5805982 0.4634924 0.6998121 0.4972940
#> [106] 0.4153063 0.2909862 0.5469323 0.6121560 0.5000731 0.4162647 0.5665244
#> [113] 0.4157386 0.5003498 0.5216893 0.5043520 0.5259336 0.4374349 0.5905698
#> [120] 0.6290429 0.4045211 0.6541578 0.4319214 0.5673303 0.5538572 0.6297819
#> [127] 0.4926927 0.3790505 0.3687178 0.5054838 0.6217428 0.4490602 0.4629686
#> [134] 0.4210480 0.4977018 0.5490171 0.5256610 0.5323598 0.6077392 0.5077406
#> [141] 0.5115820 0.6129213 0.5661409 0.3343203 0.4102714 0.4851036 0.4211769
#> [148] 0.5550966 0.4821337 0.5828197 0.6780051 0.4119590 0.3704688 0.4484376
#> [155] 0.3419939 0.5749320 0.5156625 0.4174928 0.4712343 0.6325820 0.3863714
#> [162] 0.4151049 0.6033254 0.5842202 0.5069127 0.6483842 0.4173760 0.4264816
#> [169] 0.6552875 0.4006958 0.4057597 0.7859721 0.5321690 0.7169424 0.4349792
#> [176] 0.5288682 0.5697059 0.4811072 0.4762157 0.5616735 0.5760839 0.5494225
#> [183] 0.4078381 0.5106791 0.5492321 0.5364980 0.3631797 0.5220954 0.6980757
#> [190] 0.3806673 0.5525996 0.4925890 0.6393960 0.5300620 0.5001994 0.5941849
#> [197] 0.3818391 0.6605101 0.4988376 0.6294051 0.6029637 0.6549538 0.6465341
#> [204] 0.6269271 0.4906094 0.4737846 0.4678744 0.4319913 0.7767213 0.5533653
#> [211] 0.5066803 0.5062953 0.6890323 0.6595922 0.3829071 0.6510748 0.4110026
#> [218] 0.5091385 0.4636898 0.3607468 0.4798451 0.6983867 0.3706519 0.4418332
#> [225] 0.4694361 0.5246716 0.5805739 0.4525727 0.4271134 0.4060185 0.4144180
#> [232] 0.5363131 0.5385253 0.6118266 0.6302383 0.6091166 0.5053635 0.4578109
#> [239] 0.5564455 0.7313536 0.3461167 0.3839021 0.4652478 0.5631743 0.4705506
#> [246] 0.3546107 0.3451336 0.5089301 0.4348207 0.4986162 0.3626558 0.5516112
#> [253] 0.5660044 0.3828654 0.4797113 0.5783817 0.5619527 0.5413491 0.5908250
#> [260] 0.5020984 0.4681087 0.4563402 0.4175474 0.3363384 0.5908963 0.5061498
#> [267] 0.3965401 0.6363310 0.5192099 0.5914666 0.7287171 0.5654966 0.5830175
#> [274] 0.5338142 0.4875757 0.5287558 0.4234872 0.7124835 0.5385809 0.5020249
#> [281] 0.4982946 0.5350515 0.3243927 0.3806867 0.4580532 0.5221023 0.6952034
#> [288] 0.6533267 0.4921038 0.7334513 0.5032917 0.6676040 0.7125776 0.5408955
#> [295] 0.3928002 0.3287531 0.5012287 0.3799930 0.5116441 0.6056690 0.3866806
#> [302] 0.5389672 0.6632026 0.5184388 0.5110211 0.3489454 0.5529228 0.5960493
#> [309] 0.6654337 0.4993021 0.5917869 0.6261243 0.4672141 0.3754637 0.5558017
#> [316] 0.4886690 0.6332012 0.5225871 0.4343941 0.3742135 0.4093834 0.5497649
#> [323] 0.4357071 0.4759894 0.4305244 0.7010143 0.4382944 0.4720362 0.5581313
#> [330] 0.5119253 0.6836475 0.5343088 0.6004379 0.4881342 0.4231969 0.3598382
#> [337] 0.5185657 0.5320229 0.6655590 0.3745967 0.4716455 0.5660316 0.4365982
#> [344] 0.4179904 0.4207624 0.6487914 0.5373581 0.4355690 0.4061525 0.6182296
#> [351] 0.5572310 0.4497439 0.5003683 0.7202764 0.4874284 0.4946942 0.4990330
#> [358] 0.5706586 0.4913570 0.5693478 0.6996120 0.5472478 0.4821505 0.5208240
#> [365] 0.3933438 0.4011692 0.7344256 0.6700732 0.4964954 0.7129060 0.6702195
#> [372] 0.4585901 0.6426145 0.6671890 0.4801019 0.6179605 0.5112422 0.4246498
#> [379] 0.4266410 0.5785052 0.2852388 0.4692045 0.8124865 0.4634984 0.4834419
#> [386] 0.6363941 0.5877120 0.6696918 0.5593411 0.3236144 0.4726739 0.4575294
#> [393] 0.5264682 0.5454175 0.5467156 0.6180588 0.5448598 0.5860055 0.4319808
#> [400] 0.5860403 0.5090573 0.6411664 0.4735417 0.5054485 0.5731065 0.3947812
#> [407] 0.5391848 0.5161291 0.3971941 0.4996523 0.4577200 0.4930528 0.4662944
#> [414] 0.5085744 0.4992777 0.4547934 0.4516528 0.6021815 0.8248945 0.3808702
#> [421] 0.6519603 0.6931349 0.6413948 0.4979950 0.4312137 0.4495234 0.3277589
#> [428] 0.6798028 0.5036491 0.6084424 0.4018696 0.5061163 0.5612356 0.5289293
#> [435] 0.4059210 0.4530022 0.3560057 0.5884984 0.4883997 0.3766108 0.5073782
#> [442] 0.2827512 0.6967117 0.6114084 0.5037660 0.5014076 0.5037769 0.4918257
#> [449] 0.4966520 0.5466264 0.5315695 0.4525137 0.4324339 0.4516680 0.5966435
#> [456] 0.4726886 0.6759836 0.4992247 0.3415949 0.6323136 0.5374092 0.6687381
#> [463] 0.4824828 0.3535219 0.3766057 0.3964241 0.5403719 0.5217721 0.6596528
#> [470] 0.4027495 0.4914423 0.6325325 0.4887104 0.4170075 0.4337061 0.4749667
#> [477] 0.6291675 0.4954692 0.5892452 0.5019742 0.5437775 0.4848289 0.6942602
#> [484] 0.5734855 0.6626265 0.5874782 0.5319290 0.5711304 0.5026551 0.6049298
#> [491] 0.4248816 0.6497884 0.6164713 0.5809698 0.5410465 0.6120050 0.5736361
#> [498] 0.5327275 0.5885373 0.5476825
#> 
#> $ci
#>     97.5%      2.5% 
#> 0.2297709 0.6066792 
#>